+++ banzai/calibrations.py
@@ -139,7 +139,7 @@ class CalibrationUser(Stage):
     def get_calibration_file_info(self, image):
         return dbs.cal_record_to_file_info(
             dbs.get_master_cal_record(image, self.calibration_type, self.master_selection_criteria,
-                                      self.runtime_context.db_address,
+                                      self.runtime_context.cal_db_address,
                                       use_only_older_calibrations=self.runtime_context.use_only_older_calibrations)
         )