+++ banzai/exceptions.py
@@ -4,3 +4,8 @@ class InhomogeneousSetException(Exception):
 
 class MissingCatalogException(Exception):
     pass
+
+
+class FrameNotAvailableError(Exception):
+    """Raised when a frame cannot be downloaded (NULL frameid or not in archive)"""
+    pass